backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
httpx>=0.27
PyMuPDF>=1.23
python-dotenv>=1.0
